Yemek Cesitleri ve ogun eklemeye bakıldı.

+++ DietProject.DAL/Entities/Yemek.cs
@@ -11,7 +11,7 @@ namespace DietProject.DAL.Entities
     {
         public string YemekAdi { get; set; }
         public string Aciklama { get; set; }
-        public double Kalori { get; set; }
+        public int Kalori { get; set; }
         public int KategoriId { get; set; }
         public Kategori Kategori { get; set; }
         public List<KullaniciOgunYemekPorsiyon> KullaniciOgunYemekPorsiyonlar { get; set; }
